requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
yt-dlp
google-genai